Move the `SetUpValidationMocks` method to `ControllersTest`
