+++ IntegrationTests/Controllers/ControllersTest.cs
@@ -1,5 +1,7 @@
-﻿using EUniversity.Core.Policy;
+﻿using EUniversity.Core.Models;
+using EUniversity.Core.Policy;
 using EUniversity.IntegrationTests.Mocks;
+using NSubstitute;
 
 namespace EUniversity.IntegrationTests.Controllers;
 
@@ -23,4 +25,29 @@ public class ControllersTest : IntegrationTest<MockedProgramWebApplicationFactor
         CreateAuthorizedClient(id, "student", Roles.Student);
     public HttpClient CreateTeacherClient(string? id = null) =>
         CreateAuthorizedClient(id, "teacher", Roles.Teacher);
+
+    /// <summary>
+    /// Sets up validation mocks in such a way that all
+    /// foreign key validations pass.
+    /// </summary>
+    protected void SetUpValidationMocks()
+    {
+        WebApplicationFactory.ExistenceCheckerMock
+            .ExistsAsync<AnyEntity, AnyEntityId>(Arg.Any<AnyEntityId>())
+            .ReturnsForAnyArgs(true);
+
+        WebApplicationFactory.UserManagerMock
+            .FindByIdAsync(Arg.Any<string>())
+            .Returns(new ApplicationUser()
+            {
+                Id = "test",
+                Email = "a@example.com",
+                FirstName = "Test1",
+                LastName = "Test2"
+            });
+
+        WebApplicationFactory.UserManagerMock
+            .IsInRoleAsync(Arg.Any<ApplicationUser>(), Arg.Any<string>())
+            .Returns(true);
+    }
 }
