+++ IntegrationTests/Controllers/CrudControllersTest.cs
@@ -107,29 +107,6 @@ public abstract class CrudControllersTest<TEntity, TId, TPreviewDto, TDetailsDto
     public abstract void SetUpService();
 
     /// <summary>
-    /// Sets up validation mocks in such a way that all
-    /// foreign key validations pass.
-    /// </summary>
-    protected void SetUpValidationMocks()
-    {
-        WebApplicationFactory.ExistenceCheckerMock
-            .ExistsAsync<AnyEntity, AnyEntityId>(Arg.Any<AnyEntityId>())
-            .ReturnsForAnyArgs(true);
-
-        WebApplicationFactory.UserManagerMock
-            .FindByIdAsync(Arg.Any<string>())
-            .Returns(new ApplicationUser() 
-            { 
-                Id = "test", Email = "a@example.com", 
-                FirstName = "Test1", LastName = "Test2"
-            });
-
-        WebApplicationFactory.UserManagerMock
-            .IsInRoleAsync(Arg.Any<ApplicationUser>(), Arg.Any<string>())
-            .Returns(true);
-    }
-
-    /// <summary>
     /// Route of HTTP GET for getting a page of elements, without query parameters.
     /// </summary>
     /// <remarks>
